Update controller packages to match core v42 (#6125)
Most controller packages have been updated to the versions included in
the core monorepo v42 release. The keyring controller was the only
package held back, due to the BigInt incompatibility.

The breaking change for most of these updates was related to the
removal of the `isomorphic-fetch` package. That package would polyfill
the `fetch` API if it wasn't present. It's not a breaking change for
mobile because mobile already includes its own polyfill for the `fetch`
API.

The only other breaking change is to the gas fee controller, which has
made a constructor parameter required rather than optional. This does
not affect mobile because mobile already sets that parameter..

See here for the release notes for each updated controller: https://github.com/MetaMask/core/releases/tag/v42.0.0

This progresses https://github.com/MetaMask/mobile-planning/issues/798