fix: Fix crash when switching to Linea (#6903)
* fix: Fix crash when switching to Linea

A recent `@metamask/controller-utils` update in #6874 caused the app
to crash when switching to Linea because it removed the resolution for
this package. Without that resolution, multiple copies of the package
were introduced, but the Linea patch was only applied to one of them.
The network controller was using a copy without the patch applied,
resulting in the crash.

This was resolved by restoring the resolution. We can remove it again
once we have deleted the patch.

* Update controller-utils patch to restore constants

The constants `MAINNET` and `TESTNET_TICKER_SYMBOLS` were removed in
v3.1.0, which was accidentally omitted from the changelog and not
considered in the release version (this should have been a breaking
change).

They have been restored until we no longer rely on packages that use
`@metamask/controller-utils` at versions under v3.1.0.