fix(action): octokit not supported on MetaMask repos (#7024)
* fix(action): octokit not supported on MetaMask repos

The github action was constantly failing with the following error: octokit/request-action@v2.x is not allowed to be used in MetaMask/metamask-mobile. The fix is to use a curl instead of using octokit.

* fix(action): shellcheck was failing

Shellcheck was failing in the CI. Had to craft the payload of the curl before doing the curl to fix this.