refactor: Use selectors for core network state access (#7121)
* refactor: Use selectors for core network state access

Replace direct references to NetworkController state with selectors.
Most of these examples were skipped in previous selector refactors
because they were outside of a React context, so there was no
established way to get the Redux state as input for the selector. We
can now do that by importing the Redux store directly and calling
`getState`.

This relates to https://github.com/MetaMask/mobile-planning/issues/1226

* Update WalletConnect to pass full state to selector

The WalletConnect module has been updated to pass the full Redux state
to the `selectChainId` selector, rather than just the background state.
The background state was being passed in alone because it was left
behind accidentally from an earlier version of this branch, from before
the recent store refactors/TypeScript conversion.