fix: Fix type errors in reducers (#7147)
All type errors in the reducers have been fixed. An error in the
EngineState type was exposed during this process as well, and has also
been fixed (the `PermissionControllerState` was resolving to `never`
because the wrong type parameter was passed in).

The `fiatOrders` unit tests used selectors quite a lot, so there was a
need for a mock/initial root state shape. This was introduced in the
`app/util/test` directory for use in unit tests. This initial state
object represents just the minimal initial state that has been typed so
far. After we migrate all reducers to TypeScript and fix their types,
this state object will become increasingly complete/correct.

This is a follow-up to #7110, which laid the groundwork for these type
fixes but also introduced some of them accidentally.