refactor: Update ENS utils to accept chain ID (#7150)
* refactor: Update ENS utils to accept chain ID

The ENS utility functions now accept chain ID instead of network ID.
Chain ID is easier for us to access because we know the chain ID for
all configured networks at all times, as it's a required property. The
network ID has to be looked up dynamically at runtime, meaning that we
don't know what it is until after the network has loaded.

The network ID is still used internally by the ENS utility functions
because the library we use for ENS support still expects network ID
rather than chain ID.

This was done to simplify the upcoming network controller update. The
new update replaces `network` with `networkId` and `networkStatus`,
where the ID is `null` while the network is loading. This forces us to
handle that loading state anywhere the network ID had been used.
Reducing the usages of network ID means avoiding that work.

This relates to https://github.com/MetaMask/mobile-planning/issues/1226

* Fix network type enum references

* Add `getCachedENSName` utility function

The new `getCachedENSName` function will retrieve a cached name from
the cache, ensuring that we don't need to reference the cache directly
from elsewhere in the codebase. This removes another reference to the
network ID.

The testing setup was a bit awkward because `ENSCache.cache` is a
property, which Jest offers no way to mock out at the moment. But this
is resolved in jest v29 with the `replaceProperty` method. A workaround
using hooks has been used as an interim solution.