chore: Replace local code fence transform with `@metamask/build-utils` (#7972)
## **Description**

In https://github.com/MetaMask/metamask-mobile/pull/7734, code fence
support was added to MetaMask Mobile by essentially copypasting the
implementation from the extension. This PR replaces the local code fence
implementation with a version that was extracted into its own package.

The external package implements code fence parsing and removal, as well
as linting transformed files. The consumer is responsible for
integrating the fence parser into the given build system, and
configuring the ESLint instance used for linting. Consequently, only the
Metro transform and the ESLint instance initialization remain in this
repository.

## **Related issues**

- https://github.com/MetaMask/core/pull/3577
- https://github.com/MetaMask/metamask-extension/pull/22033

## **Manual testing steps**

1. Check that different build types are built correctly