fix: substr undefined when cancelling a legacy transaction (#8274)
## **Description**

This PR aims to fix an issue when cancelling a legacy transaction
parsing the `data` property when the property is `undefined`. The call
for parsing comes from the gasPooling which checks it periodically.

<!--
Write a short description of the changes included in this pull request,
also include relevant motivation and context. Have in mind the following
questions:
1. What is the reason for the change?
2. What is the improvement/solution?
-->

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-mobile/issues/7149